Change FiberTrace coefficients from double to float
As the tests have shown, a FiberTrace reconstructed from
the persisted FiberTrace showed an offset in one row by one
pixel, caused by persisting the double FiberTrace function
coefficients as floats. changing the coefficients from double
to float seems to solve the problem. If not, we'll need to change
back to doubles and persist doubles in pfsFiberTrace.

+++ include/pfs/drp/stella/FiberTraces.h
@@ -176,7 +176,7 @@ class FiberTrace {
     bool isFiberTraceProfileFittingControlSet() const {return _isFiberTraceProfileFittingControlSet;}
     size_t getWidth() const {return _trace->getImage()->getWidth();}
     size_t getHeight() const {return _trace->getImage()->getHeight();}
-    ndarray::Array<double, 1, 1> getTraceCoefficients() const;
+    ndarray::Array<float, 1, 1> getTraceCoefficients() const;
 //    bool setTraceCoefficients(ndarray::Array<double, 1, 1> const& coeffs);
     PTR(FiberTrace) getPointer();
 
@@ -320,9 +320,9 @@ namespace math{
                                                                      const PTR(const FiberTraceFunctionFindingControl) &fiberTraceFunctionFindingControl);
   
   struct FindCenterPositionsOneTraceResult{
-      std::vector<double> apertureCenterIndex;/// CONVERT ALL TO FLOAT???
-      std::vector<double> apertureCenterPos;
-      std::vector<double> eApertureCenterPos;
+      std::vector<float> apertureCenterIndex;/// CONVERT ALL TO FLOAT???
+      std::vector<float> apertureCenterPos;
+      std::vector<float> eApertureCenterPos;
   };
   
   /**
